pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pearlnn"
version = "0.1.0"
description = "Parameter Extraction And Reverse Learning Neural Network"
authors = [{name = "PEARLNN Community"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["electronics", "neural-network", "parameter-fitting", "bayesian", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "opencv-python>=4.5.0",
    "Pillow>=9.0.0",
    "scipy>=1.7.0",
    "requests>=2.25.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
    "rich>=12.0.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.910",
    "pre-commit>=2.0.0",
    "jupyter>=1.0.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=8.0.0",
    "mkdocstrings>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/pearlnn/pearlnn"
Documentation = "https://pearlnn.readthedocs.io"
Repository = "https://github.com/pearlnn/pearlnn"
Issues = "https://github.com/pearlnn/pearlnn/issues"
"Community" = "https://github.com/pearlnn/pearlnn/discussions"

[project.scripts]
pearlnn = "pearlnn.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--verbose --color=yes"
testpaths = ["pearlnn/tests"]
python_files = ["test_*.py"]